# Contenido_API/bin/src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  description: "Esta API REST se encarga de manejar la información de la base de datos\
    \ del microservicio de Contenidos del servicio de Netflix. Proporciona varios\
    \ endpoints para gestionar películas, series y actores, permitiendo realizar operaciones\
    \ CRUD (Crear, Leer, Actualizar, Eliminar) sobre los recursos relacionados. Esta\
    \ API facilita la interacción entre el frontend y el backend del sistema, asegurando\
    \ que los datos sobre el catálogo de contenido estén siempre actualizados y accesibles\
    \ para los usuarios y administradores del servicio."
  title: API de Contenido
  version: "1.0"
servers:
- url: https://api.server/v1/content/
...
paths:
...
  /films/actor/{actorID}:
    get:
      description: Devuelve una lista de películas filtradas por el id del actor que
        participa en ellas
      parameters:
      - description: El ID del actor por el cual se filtrarán las películas.
        explode: false
        in: path
        name: actorID
        required: true
        schema:
          example: 1
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/film'
                type: array
          description: Lista de películas filtrada por actor obtenida con éxito
        "404":
          description: (NotFound) El recurso solicitado no existe en la base de datos.
        "400":
          description: (BadRequest) Los datos enviados no son correctos o faltan datos
            obligatorios en la solicitud.
      summary: Lista películas por ID de actor
      tags:
      - films
      x-accepts:
      - application/json
      x-tags:
      - tag: films
...
  /series/actor/{actorID}:
    get:
      description: Devuelve una lista de series filtradas por el id del actor que
        participa en ellas
      parameters:
      - description: El ID del actor por el cual se filtrarán las series.
        explode: false
        in: path
        name: actorID
        required: true
        schema:
          example: 1
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/serie'
                type: array
          description: Lista de series filtrada por actor obtenida con éxito
        "404":
          description: (NotFound) El recurso solicitado no existe en la base de datos.
        "400":
          description: (BadRequest) Los datos enviados no son correctos o faltan datos
            obligatorios en la solicitud.
      summary: Lista series por ID de actor
      tags:
      - series
      x-accepts:
      - application/json
      x-tags:
      - tag: series
...
  /actors/{id}:
    delete:
      description: Elimina un actor de la base de datos. El ID del actor debe proporcionarse
        en la URL para identificar al actor que será eliminado.
      parameters:
      - description: El ID del actor a eliminar
        explode: false
        in: path
        name: id
        required: true
        schema:
          example: 1
          type: integer
        style: simple
      responses:
        "204":
          description: Actor eliminado con éxito
        "404":
          description: (NotFound) El recurso solicitado no existe en la base de datos.
        "400":
          description: (BadRequest) Los datos enviados no son correctos o faltan datos
            obligatorios en la solicitud.
      summary: Elimina un actor
      tags:
      - actors
      x-accepts:
      - application/json
      x-tags:
      - tag: actors
    get:
      description: "Devuelve los detalles de un actor específico a partir de su ID,\
        \ como su nombre, fecha de nacimiento y las imágenes asociadas."
      parameters:
      - description: El ID del actor a obtener
        explode: false
        in: path
        name: id
        required: true
        schema:
          example: 1
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/actor'
          description: Actor obtenido con éxito
        "404":
          description: (NotFound) El recurso solicitado no existe en la base de datos.
      summary: Devuelve los detalles de un actor específico
      tags:
      - actors
      x-accepts:
      - application/json
      x-tags:
      - tag: actors
    put:
      description: "Actualiza los detalles de un actor existente. Este endpoint permite\
        \ modificar la información de un actor, incluyendo su nombre o fecha de nacimiento."
      parameters:
      - description: El ID del actor a actualizar
        explode: false
        in: path
        name: id
        required: true
        schema:
          example: 1
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/actor'
        description: Objeto actor con nuevos valores
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/actor'
          description: Actor actualizado con éxito
        "404":
          description: (NotFound) El recurso solicitado no existe en la base de datos.
        "400":
          description: (BadRequest) Los datos enviados no son correctos o faltan datos
            obligatorios en la solicitud.
      summary: Actualiza un actor
      tags:
      - actors
      x-content-type: application/json
      x-accepts:
      - application/json
      x-tags:
      - tag: actors
components:
...
  schemas:
    film:
      example:
        genreID: 1
        duration: 120
        ArrayActors:
        - 101
        - 102
        - 103
        photoUrls:
        - photoUrls
        - photoUrls
        release_year: 2023
        description: Film Sinopsis
        id: 1
        title: Film Title
      properties:
        id:
          example: 1
          type: integer
        title:
          example: Film Title
          type: string
        genreID:
          example: 1
          type: integer
        release_year:
          example: 2023
          type: integer
        duration:
          example: 120
          type: integer
        description:
          example: Film Sinopsis
          type: string
        photoUrls:
          items:
            type: string
            xml:
              name: photoUrl
          type: array
          xml:
            wrapped: true
        ArrayActors:
          example:
          - 101
          - 102
          - 103
          items:
            type: integer
          type: array
      required:
      - duration
      - genreID
      - id
      - release_year
      - title
      type: object
    serie:
      example:
        genreID: 1
        ArrayActors:
        - 101
        - 102
        - 103
        seasons: 5
        photoUrls:
        - photoUrls
        - photoUrls
        release_year: 2022
        description: Series Sinopsis
        id: 1
        title: Series Title
      properties:
        id:
          example: 1
          type: integer
        title:
          example: Series Title
          type: string
        seasons:
          example: 5
          type: integer
        release_year:
          example: 2022
          type: integer
        genreID:
          example: 1
          type: integer
        description:
          example: Series Sinopsis
          type: string
        photoUrls:
          items:
            type: string
            xml:
              name: photoUrl
          type: array
          xml:
            wrapped: true
        ArrayActors:
          example:
          - 101
          - 102
          - 103
          items:
            type: integer
          type: array
      required:
      - genreID
      - id
      - seasons
      - title
      type: object
    actor:
      example:
        photoUrls:
        - photoUrls
        - photoUrls
        birthday_date: 2003-07-16
        name: Actor Name
        id: 1
      properties:
        id:
          example: 1
          type: integer
        name:
          example: Actor Name
          type: string
        birthday_date:
          example: 2003-07-16
          format: date-time
        photoUrls:
          items:
            type: string
            xml:
              name: photoUrl
          type: array
          xml:
            wrapped: true
      required:
      - birthday_date
      - id
      - name
      type: object
...